freelist: inline production operation delegates
